refactor(form-builder): extract form builder logic into custom hook and components
Split form builder logic into reusable useFormBuilder hook and separate components
Add error handling improvements with development logging
Clean up event actions by removing unused code and using error handler
