+++ app/shared/lib/error-handler/index.ts
@@ -4,6 +4,10 @@ import type { ActionData } from "../../types";
 
 export async function handleServiceError(error: unknown): Promise<ActionData> {
   if (error instanceof Prisma.PrismaClientKnownRequestError) {
+    if (process.env.NODE_ENV === "development") {
+      console.error("Prisma error:", error);
+    }
+
     return {
       success: false,
       message: clientPrismaErrors[error.code] || clientPrismaErrors.DEFAULT,
@@ -12,6 +16,10 @@ export async function handleServiceError(error: unknown): Promise<ActionData> {
   }
 
   if (error instanceof Error) {
+    if (process.env.NODE_ENV === "development") {
+      console.error("Error:", error);
+    }
+
     return {
       success: false,
       message: error.message || clientPrismaErrors.DEFAULT,
